+++ config/hungarian-zip-codes.php
@@ -4,5 +4,10 @@ return [
     /*
      * You can specify the table name of the zip codes
      */
-    'table_name' => 'hungarian_zip_codes'
+    'table_name' => 'hungarian_zip_codes',
+
+    /*
+     * You can save the Budapest's districts to separate field
+     */
+    'district' => true
 ];