+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to `laravel-hungarian-zip-codes` will be documented in this file.
 
+## 1.0.3 - 2022-02-05
+
+- Fix php version in composer file
+- Seeders publish fix
+- Renaming the migration file
+- Fix the table config name in migration file
+- Added the data json file to the publishable files
+- Seeder script refactor and fixes
+- Migration file timestamps deleted
+
 ## 1.0.2 - 2022-02-05
 
 - Readme license and changelog files are added